requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
numpy==1.24.3
pyarrow==14.0.1
boto3==1.29.7
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
xgboost==2.0.2
scikit-learn==1.3.2
tensorflow==2.14.0
yfinance==0.2.32
python-dotenv==1.0.0
pydantic==2.5.0
redis==5.0.1
httpx==0.25.1
asyncio==3.4.3
python-multipart==0.0.6
prometheus-client==0.19.0
optuna==3.4.0
joblib==1.3.2